更新 Cargo.toml 以升级 mysql 和 thiserror 依赖，新增 tempfile 作为开发依赖。修改 DEVELOPMENT.md 文档，增强 CLI 输入解析功能，支持多种输入格式并提供友好的错误提示。重构 CLI 解析逻辑，新增 ResolvedTest 结构体以支持智能解析。添加多个单元测试以验证新功能的正确性，提升代码质量和稳定性。
