+++ Cargo.toml
@@ -13,14 +13,14 @@ path = "src/lib.rs"
 
 [dependencies]
 # MySQL 驱动
-mysql = "24.0"
+mysql = "26.0.0"
 
 # CLI 解析
 clap = { version = "4.0", features = ["derive"] }
 
 # 错误处理
 anyhow = "1.0"
-thiserror = "1.0"
+thiserror = "2.0.12"
 
 # 日志
 log = "0.4"
@@ -55,6 +55,9 @@ evalexpr = "12.0.2"
 # Pest parser (可选)
 pest = { version = "2.7", optional = true }
 pest_derive = { version = "2.7", optional = true }
+
+[dev-dependencies]
+tempfile = "3.0"
 
 [features]
 default = ["pest"]
