+++ DEVELOPMENT.md
@@ -25,13 +25,20 @@
    cargo build --release
    ```
 
-4. **运行单个测试文件**
+4. **运行测试文件（支持多种格式）**
    ```bash
-   # 第一次：录制结果文件（生成 r/basic.result）
+   # 单个测试名称
    cargo run -- --record basic
-
-   # 第二次：与预期结果进行比对
    cargo run -- basic
+
+   # 目录下所有测试
+   cargo run -- t/features
+
+   # 混合多种格式
+   cargo run -- basic advanced.test t/regression
+
+   # 部分匹配（匹配所有包含 "user" 的测试）
+   cargo run -- user
    ```
 
 5. **批量运行全部测试**
@@ -137,6 +144,12 @@
   - [x] 测试选项参数
   - [x] 邮件配置参数
   - [x] 参数验证逻辑
+- [x] **增强 CLI 输入解析** ⭐**新增 (2025-01-17)**
+  - [x] 支持多种输入格式：测试名称、目录、文件路径、部分匹配
+  - [x] 实现 `ResolvedTest` 结构体和智能解析逻辑
+  - [x] 自动去重和一致性排序
+  - [x] 友好的错误提示和使用建议
+  - [x] 完整的测试覆盖 (5个测试用例全部通过)
 
 ### ✅ 已完成阶段
 
@@ -286,6 +299,19 @@
 - `default_parser()`: 自动选择可用的最佳实现
 - `create_parser(type)`: 支持显式指定解析器类型
 
+### 9. CLI 输入解析策略 ⭐**新增 (2025-01-17)**
+**决策**: 实现智能输入解析系统，支持多种格式
+**原因**:
+- **用户体验**: 提供灵活的输入方式，减少使用复杂度
+- **向后兼容**: 保持与原有测试名称格式的兼容性
+- **扩展性**: 支持目录、文件路径、部分匹配等高级功能
+- **健壮性**: 提供友好的错误提示和使用建议
+**实现**:
+- `ResolvedTest` 结构体：统一的测试表示
+- 多策略解析：按优先级尝试不同解析方式
+- 智能去重：自动移除重复的测试文件
+- 错误引导：提供具体的使用格式建议
+
 ## 代码质量指标
 
 - **编译状态**: ✅ 通过
@@ -337,6 +363,12 @@
   - 测试变量设置和获取
   - 测试递归展开
   - 测试 let 语句解析
+- ✅ **CLI 输入解析测试** (`tests/input_resolution_test.rs`) ⭐**新增**
+  - 测试多种输入格式解析
+  - 测试目录遍历和文件发现
+  - 测试部分匹配和错误处理
+  - 测试去重和排序逻辑
+  - 5个测试用例全部通过
 - 🔄 命令处理器测试 (`tests/unit/handlers/`)
   - 测试各个命令处理器
   - 测试命令注册机制
@@ -399,21 +431,23 @@
 | **项目的长期规划是什么？** | 详见上文"后续开发计划"，包括并发支持、XML 报告和完整的命令集实现。 |
 | **Pest Parser vs 手写 Parser 有什么区别？** | Pest Parser 是基于正式语法定义的结构化实现，更易维护和扩展；手写 Parser 是原有实现，经过充分测试。两者功能完全等价，可通过 feature flag 切换。 |
 | **为什么 Pest Parser 是默认的？** | Pest 提供更清晰的语法定义、更好的错误报告和更易维护的代码结构，同时与手写 parser 保持 100% 兼容性。 |
+| **CLI支持哪些输入格式？** | 支持测试名称(`basic`)、文件名(`basic.test`)、目录(`t/features`)、文件路径(`path/to/test.test`)、部分匹配(`user`)和混合格式，具有智能去重和友好错误提示。 |
+| **如何运行目录下所有测试？** | 使用 `cargo run -- t/目录名` 可以运行指定目录下的所有 `.test` 文件，会自动递归搜索并按名称排序。 |
 
 ---
 
-**最后更新**: 2025年01月16日  
-**当前版本**: v0.2.1-dev  
+**最后更新**: 2025年01月17日  
+**当前版本**: v0.2.2-dev  
 **开发者**: [项目团队]
 
-### 📋 最新进展总结 (2025-01-16)
+### 📋 最新进展总结 (2025-01-17)
 
-本次更新主要完成了 **Pest Parser 的关键修复**：
+本次更新主要完成了 **CLI 增强和输入解析优化**：
 
-- 🔧 **解决了 Pest Parser 的核心问题**：大小写敏感性、多行 SQL 合并、变量解析等
-- ✅ **测试验证通过**：主要功能测试正常，仅有细微格式差异待完善  
-- 🎯 **双解析器架构稳定**：Pest Parser 为默认选择，手写解析器作为可靠备选
-- 📈 **项目稳定性提升**：解析层现已基本稳定，为后续功能开发奠定基础
+- 🚀 **CLI 功能大幅增强**：支持多种输入格式（测试名称、目录、文件路径、部分匹配）
+- ✅ **智能解析系统**：实现 `ResolvedTest` 结构体和多策略解析逻辑
+- 🛡️ **健壮性提升**：自动去重、友好错误提示、完整测试覆盖
+- 📈 **用户体验改善**：大幅简化了命令行使用复杂度，提供直观的操作方式
 
 ## 🎉 最新功能亮点
