--- r/expected_error_test.result
@@ -1,1 +1,0 @@
-Got one of the listed errors
