+++ r/test_line_numbers.result
@@ -1,0 +1,5 @@
+SELECT 1
+1
+"Line 5"
+SELECT 2
+2
