+++ src/tester/pest_parser.rs
@@ -43,9 +43,12 @@ impl PestParser {
     fn convert_to_queries(&mut self, pairs: pest::iterators::Pairs<Rule>) -> Result<Vec<Query>> {
         let mut queries = Vec::new();
         let mut pending_sql_lines = Vec::new();
-        let line_num = 1;
+        let mut line_num = 1;
 
         for pair in pairs {
+            // Update line number based on the pair's position
+            let pair_line = pair.line_col().0;
+            line_num = pair_line;
             match pair.as_rule() {
                 Rule::test_file => {
                     // Recursively process the test file contents
@@ -381,6 +384,79 @@ impl QueryParser for PestParser {
             .map_err(|e| anyhow!("Pest parsing error: {}", e))?;
         
         self.convert_to_queries(pairs)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::tester::parser::{create_parser};
+    use crate::tester::query::QueryType;
+
+    #[test]
+    fn test_pest_parse_simple_query() {
+        let mut parser = create_parser("pest").expect("Failed to create pest parser");
+        let content = "SELECT 1;";
+        let queries = parser.parse(content).expect("Failed to parse simple query");
+        
+        assert_eq!(queries.len(), 1);
+        assert_eq!(queries[0].query_type, QueryType::Query);
+        assert_eq!(queries[0].query.trim(), "SELECT 1");
+    }
+
+    #[test]
+    fn test_pest_parse_command() {
+        let mut parser = create_parser("pest").expect("Failed to create pest parser");
+        let content = "--echo hello world";
+        let queries = parser.parse(content).expect("Failed to parse command");
+        
+        assert_eq!(queries.len(), 1);
+        assert_eq!(queries[0].query_type, QueryType::Echo);
+        assert_eq!(queries[0].query, "hello world");
+    }
+
+    #[test]
+    fn test_pest_parse_comment() {
+        let mut parser = create_parser("pest").expect("Failed to create pest parser");
+        let content = "# This is a comment";
+        let queries = parser.parse(content).expect("Failed to parse comment");
+        
+        assert_eq!(queries.len(), 1);
+        assert_eq!(queries[0].query_type, QueryType::Comment);
+        assert!(queries[0].query.contains("This is a comment"));
+    }
+
+    #[test]
+    fn test_pest_parse_if_statement() {
+        let mut parser = create_parser("pest").expect("Failed to create pest parser");
+        let content = "if ($var > 0) {\n--echo positive\n}";
+        let queries = parser.parse(content).expect("Failed to parse if statement");
+        
+        // Should have at least the if statement
+        assert!(!queries.is_empty());
+        let if_query = queries.iter().find(|q| q.query_type == QueryType::If);
+        assert!(if_query.is_some());
+        assert_eq!(if_query.unwrap().query.trim(), "$var > 0");
+    }
+
+    #[test]
+    fn test_pest_parse_delimiter_change() {
+        let mut parser = create_parser("pest").expect("Failed to create pest parser");
+        let content = "--delimiter //\nSELECT 1//";
+        let queries = parser.parse(content).expect("Failed to parse delimiter change");
+        
+        assert!(queries.len() >= 2);
+        
+        // Find delimiter command
+        let delimiter_query = queries.iter().find(|q| q.query_type == QueryType::Delimiter);
+        assert!(delimiter_query.is_some());
+        assert_eq!(delimiter_query.unwrap().query, "//");
+        
+        // Find SQL query
+        let sql_query = queries.iter().find(|q| q.query_type == QueryType::Query);
+        assert!(sql_query.is_some());
+        // The query may still contain the delimiter when parsed
+        assert!(sql_query.unwrap().query.contains("SELECT 1"));
     }
 }
