--- t/expected_error_test.test
@@ -1,3 +1,0 @@
---disable_query_log
---error 0
-SELECT * FROM non_existing_table;
