--- tests/unit/mod.rs
@@ -1,10 +1,0 @@
-pub mod connection_manager;
-pub mod database;
-pub mod error_handler;
-pub mod error_utils;
-pub mod handlers;
-pub mod loader;
-pub mod parser;
-#[cfg(feature = "pest")]
-pub mod pest_parser;
-pub mod tester;
